refactor: add separate stacks for Backend and Frontend
