+++ client/src/constants/apiPaths.ts
@@ -1,5 +1,5 @@
 const API_PATHS = {
-  product: "https://tqepc8bhxl.execute-api.us-east-2.amazonaws.com/dev",
+  product: "https://j5nyxrx3vg.execute-api.us-east-2.amazonaws.com/dev",
   order: "https://.execute-api.eu-west-1.amazonaws.com/dev",
   import: "https://.execute-api.eu-west-1.amazonaws.com/dev",
   bff: "https://.execute-api.eu-west-1.amazonaws.com/dev",
