+++ deploy-cdk/stacks/backend/index.ts
@@ -1,0 +1,3 @@
+import { BackendStack } from "./backend-stack";
+
+export default BackendStack;
