+++ deploy-cdk/stacks/frontend/index.ts
@@ -1,0 +1,3 @@
+import { FrontendStack } from "./frontend-stack";
+
+export default FrontendStack;
