+++ server/package.json
@@ -4,7 +4,8 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "test": "jest"
+    "test": "jest",
+    "deploy": "cd ../deploy-cdk && npm run build && cdk deploy BackendStack"
   },
   "keywords": [],
   "author": "",
